move strings to resources: Controls dir
